+++ readme.md
@@ -1,12 +1,24 @@
-# JavaScript Arithmetic Lab
+# Arithmetic Operators Lab
 
-This pCoding Lab is a simple JavaScript Arithmetic Lab designed to help you practice basic and advanced arithmetic operations in JavaScript. The lab includes functions for addition, subtraction, multiplication, division, modulus, and exponentiation.
+# Description
+This lab is designed to help you practice basic and advanced arithmetic operations in JavaScript.
 
-## Usage
+## Completing the Lab
 
-Open the `script.js` file, and make necessary modifications for the test cases to pass.
+Open the `arithmetic-functions.js` file and ensure the below tasks are completed so that the test cases pass. The file can be found under the `src/main/java/com/revature` package.
 
+Tasks to Complete:
+[ ] Implement the add() function
+[ ] Implement the subtract() function
+[ ] Implement the multiply() function
+[ ] Implement the divide() function
+[ ] Implement the modulus() function
+[ ] Implement the exponentiate() function
 
 ## Running Tests
 
-To run the tests, click on the `test beaker` icon and run the test cases.
+- To run the tests, click on the `test beaker` icon and run the test cases.
+- To run the program for manual testing, you can call any of the functions you've implemented. Call these functions within the `index.js` file's App object's `start()` method. Run the program using the following command: `node src/main/java/com/revature/index.js` 
+
+Below is an example of manually testing the `add()` function:
+![example manual testing image](/images/man-1.png)
