+++ src/main/java/com/revature/arithmetic-functions.js
@@ -1,0 +1,80 @@
+/**
+ * The add function adds two numbers together and returns the sum.
+ *
+ * @param {Number} a - A numeric value
+ * @param {Number} b - Another numeric value that should be added to 'a'
+ * @returns {Number} - The sum of the values within the parameters 'a' and 'b'
+ */
+function add(a, b) {
+    // TODO: Implement the add function
+    return "replace me";
+}
+
+/**
+ * The subtract function subtracts a number from the other and returns the difference.
+ *
+ * @param {Number} a - A numeric value
+ * @param {Number} b - Another numeric value that should be subtracted from 'a'
+ * @returns {Number} - The difference of the values within the parameters 'a' and 'b'
+ */
+function subtract(a, b) {
+    // TODO: Implement the subtract function
+    return "replace me";
+}
+
+/**
+ * The multiply function multiplies two numbers and returns the product.
+ *
+ * @param {Number} a - A numeric value
+ * @param {Number} b - Another numeric value that should be multiplied with 'a'
+ * @returns {Number} - The product of the values within the parameters 'a' and 'b'
+ */
+function multiply(a, b) {
+    // TODO: Implement the multiply function
+    return "replace me";
+}
+
+/**
+ * The divide function divides one number from another and returns the quotient. This function should also ensure that if either value is 0, the value undefined should be returned.
+ *
+ * @param {Number} a - A numeric value
+ * @param {Number} b - Another numeric value that 'a' should be divided by
+ * @returns {Number} - The quotient of the values within the parameters 'a' and 'b'
+ */
+function divide(a, b) {
+    // TODO: Implement the divide function
+    return "replace me";
+}
+
+/**
+ * The modulus function divides one number from another and returns the remainder.
+ *
+ * @param {Number} a - A numeric value
+ * @param {Number} b - Another numeric value that 'a' should be divided by
+ * @returns {Number} - The remainder of the values within the parameters 'a' and 'b'
+ */
+function modulus(a, b) {
+    // TODO: Implement the modulus function
+    return "replace me";
+}
+
+/**
+ * The exponentiate function returns the power of the base and its exponent.
+ *
+ * @param {Number} base - A numeric value
+ * @param {Number} exponent - Another numeric value that represents how many times the base should be multipled by itself
+ * @returns {Number} - The power of the values within the parameters 'a' and 'b'
+ */
+function exponentiate(base, exponent) {
+    // TODO: Implement the exponentiate function
+    return "replace me";
+}
+
+// exporting allows index.js to call these functions
+exports.add = add;
+exports.subtract = subtract;
+exports.multiply = multiply;
+exports.divide =divide;
+exports.modulus = modulus;
+exports.exponentiate =exponentiate;
+
