+++ src/main/java/com/revature/index.html
@@ -3,10 +3,11 @@
 <head>
     <meta charset="UTF-8">
     <title>JavaScript Arithmetic Lab</title>
-    <script src="script.js"></script>
 </head>
 <body>
     <h1>JavaScript Arithmetic Lab</h1>
-    <div id="output"></div>
+    <p>No need to view or edit this file. This file is necessary for testing.</p>
+    
+    <script src="arithmetic-functions.js"></script>
 </body>
 </html>
