+++ src/main/java/com/revature/index.js
@@ -1,0 +1,17 @@
+// import functions from arithmetic-functions.js for manual testing
+const {add, subtract, multiply, modulus, divide, exponentiate} = require("./arithmetic-functions.js")
+
+/**
+ * The main application object.
+ */
+const App = {
+    /**
+    * The below function starts the application.
+    */
+    start: function () {
+        // Optional: add function calls here for manual testing
+    }
+};
+
+// Start the application when the DOM is fully loaded
+App.start();
