--- src/main/java/com/revature/script.js
@@ -1,29 +1,0 @@
-function add(a, b) {
-    // TODO: Implement the addition function
-}
-
-function subtract(a, b) {
-    // TODO: Implement the subtraction function
-}
-
-function multiply(a, b) {
-    // TODO: Implement the multiplication function
-}
-
-function divide(a, b) {
-    // TODO: Implement the division function, also make sure to check if any of the values are = 0 to prevent any errors!
-    // Remeber Java doesn't have the notion of an "undefined" state, if you try returning undefined it'll convert it to null! 
-
-
-}
-
-function modulus(a, b) {
-    // TODO: Implement the modulus function
-}
-
-function exponentiate(base, exponent) {
-    // TODO: Implement the exponentiation function
-}
-
-
-
